opt: Move data-driven testing framework to separate package
The SQL data-driven testing framework is useful on its own,
with minimal dependencies. This PR separates it into its own
package so that tests can import it without importing a chain
of other Cockroach packages (like various SQL packages).
Optgen tests run much faster now that they're not transitively
importing a large chunk of Cockroach just to use data-driven
testing.

Release note: None
