+++ pkg/sql/opt/optgen/lang/scanner_test.go
@@ -23,11 +23,11 @@ import (
 	"testing"
 	"testing/iotest"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/testutils"
+	"github.com/cockroachdb/cockroach/pkg/sql/opt/testutils/datadriven"
 )
 
 func TestScanner(t *testing.T) {
-	testutils.RunDataDrivenTest(t, "testdata/scanner", func(d *testutils.TestData) string {
+	datadriven.RunTest(t, "testdata/scanner", func(d *datadriven.TestData) string {
 		// Only scan command supported.
 		if d.Cmd != "scan" {
 			t.FailNow()
